style(rage): explicit casts on 4 header-wide C4244 narrowing sites
RageUtil.h / RageTimer.h are included by nearly every translation unit,
so these 4 sites accounted for the bulk of the C4244/C4267 raw warning
volume (backlog item 2) despite being 4 lines. All four are intentional,
safe narrowing -- adding static_cast documents that instead of relying
on the /wd4244 /wd4267 suppression:

- RageTimer::GetTimeSinceStartFast(): double -> float (the "fast",
  lower-precision sibling of GetTimeSinceStart()).
- MersenneTwister ctor: time_t -> mt19937::result_type (uint32) --
  standard "seed from the low bits of the clock" idiom.
- RandomFloat(): uniform_real_distribution<>'s double result -> float.
- FindIndex(): iterator difference (ptrdiff_t) -> int return type.

A full clean Release rebuild of the StepMania target (459 files) now
shows 0 C4244/C4267 -- that build had no other sites. A Debug/WITH_TESTS
build (sm_engine, which the tests + Debug config surface more of, per
the same "Debug shows what Release's optimizer folds away" pattern from
item 12) still has ~297 unique sites across ~80 files; /wd4244/4267
stays in src/CMakeLists.txt until those are triaged file-by-file.
Backlog item 2 updated with the real counts (the prior "~4.4k" figure
was stale/uncounted).

Verified: sm_tests 5966/226 (Debug), Release StepMania-R.exe builds.

+++ src/RageTimer.h
@@ -24,7 +24,7 @@ public:
 	float PeekDeltaTime() const { return Ago(); }
 
 	static double GetTimeSinceStart( bool bAccurate = true );	// seconds since the program was started
-	static float GetTimeSinceStartFast() { return GetTimeSinceStart(false); }
+	static float GetTimeSinceStartFast() { return static_cast<float>( GetTimeSinceStart(false) ); }
 	static std::uint64_t GetUsecsSinceStart();
 
 	/* Get a timer representing half of the time ago as this one. */
